+++ test/experiment/trending.dart
@@ -1,9 +1,0 @@
-import "package:github/server.dart";
-
-void main() {
-  var github = createGitHubClient();
-
-  github.explore
-      .listTrendingRepositories(language: "Dart", since: "month")
-      .listen((repo) => print("${repo.title}: ${repo.description}"));
-}